fix(renderer): lazy-load bug report surface
